Did corrections after code review.

+++ src/main/java/frc/robot/interfaces/LEDStripInterface.java
@@ -4,7 +4,13 @@ import frc.robot.lib.LEDColour;
 public interface LEDStripInterface {
     public void setColour(LEDColour c);
 
-    public void setProgressColour(LEDColour c1, LEDColour c2, int percent);
+    /**
+     * Fills the led strip like a progress bar.
+     * @param c1 Colour for completed area
+     * @param c2 Colour for unfinished area
+     * @param percent Percet of leds to fill from 0-1
+     */
+    public void setProgressColour(LEDColour c1, LEDColour c2, double percent);
 
     public void setIdle();
 }
